+++ app/src/main/java/com/grigorev/diploma/util/Converters.kt
@@ -12,10 +12,10 @@ class Converters {
     fun fromAttachmentType(value: AttachmentType) = value.name
 
     @TypeConverter
-    fun fromSet(set: Set<Long>): String = set.joinToString("-")
+    fun fromSet(set: Set<Int>): String = set.joinToString("-")
 
     @TypeConverter
-    fun toSet(data: String): Set<Long> =
+    fun toSet(data: String): Set<Int> =
         if (data.isBlank()) emptySet()
-        else data.split("-").map { it.toLong() }.toSet()
+        else data.split("-").map { it.toInt() }.toSet()
 }
